PRESS4-360 | Added paypal event to commerce
