+++ includes/Helpers/Plugin.php
