+++ includes/SiteCapabilities.php
