+++ includes/SubscriberInterface.php
