--- src/Listeners/Yith.php
@@ -1,46 +1,0 @@
-<?php
-
-namespace NewfoldLabs\WP\Module\Data\Listeners;
-
-/**
- * Monitors Yith events
- */
-class Yith extends Listener {
-
-	/**
-	 * Register the hooks for the listener
-	 *
-	 * @return void
-	 */
-	public function register_hooks() {
-		// Paypal Connection
-		add_filter( 'pre_update_option_yith_ppwc_merchant_data_production', array( $this, 'paypal_connection' ), 10, 2 );
-	}
-
-	/**
-	 * PayPal connected
-	 *
-	 * @param string $new_option New value of the yith_ppwc_merchant_data_production option
-	 * @param string $old_option Old value of the yith_ppwc_merchant_data_production option
-	 *
-	 * @return string The new option value
-	 */
-	public function paypal_connection( $new_option, $old_option ) {
-		$url =  is_ssl() ? "https://" : "http://"; 
-		$url .= $_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI'];
-		$data = array( 
-			"label_key" => "provider",
-			"provider" 	=> "yith_paypal",
-			"page" 		=> $url
-		);
-		if ( $new_option !== $old_option && ! empty( $new_option ) ) {	
-			$this->push(
-				"payment_connected",
-				$data
-			);
-		}
-
-		return $new_option;
-	}
-
-}
